requirements.txt
Flask==2.0.1
Flask_Login==0.5.0
Flask_WTF==1.0.0
matplotlib==3.5.0
numpy==1.21.2
pandas==1.3.4
pdfkit==1.0.0
PyMySQL==1.0.2
PyYAML==6.0
scikit_learn==1.0.2
scipy==1.7.1
seaborn==0.11.2
Werkzeug==2.0.2
WTForms==3.0.1
